chore: refactor sharing intent logic
